refactor: move item creation form handler
